backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets>=13.0
opencv-python-headless==4.10.0.84
torch>=2.9.0,<3.0.0
torchvision>=0.24.0,<0.26.0
numpy==1.26.4
python-multipart==0.0.9
supabase==2.7.4
python-dotenv==1.0.1
Pillow==10.4.0
supervision==0.22.0
groq==0.11.0
aiofiles==23.2.1